Move AnmVm to its own file and fixup a bit AnmManager constructor
