+++ src/AnmManager.cpp
@@ -3,36 +3,7 @@
 #include "GameErrorContext.hpp"
 #include "Supervisor.hpp"
 #include "i18n.hpp"
-
-AnmVm::AnmVm()
-{
-    this->spriteNumber = -1;
-}
-
-void AnmVm::Initialize()
-{
-    this->uvScrollPos.y = 0.0;
-    this->uvScrollPos.x = 0.0;
-    this->scaleInterpFinalX = 0.0;
-    this->scaleInterpFinalY = 0.0;
-    this->angleVel.z = 0.0;
-    this->angleVel.y = 0.0;
-    this->angleVel.x = 0.0;
-    this->rotation.z = 0.0;
-    this->rotation.y = 0.0;
-    this->rotation.x = 0.0;
-    this->scaleX = 1.0;
-    this->scaleY = 1.0;
-    this->scaleInterpEndTime = 0;
-    this->alphaInterpEndTime = 0;
-    this->color = D3DCOLOR_RGBA(0xff, 0xff, 0xff, 0xff);
-    D3DXMatrixIdentity(&this->matrix);
-    this->flags = AnmVmFlags_0 | AnmVmFlags_1;
-    this->autoRotate = 0;
-    this->pendingInterrupt = 0;
-    this->posInterpEndTime = 0;
-    this->currentTimeInScript.Initialize();
-}
+#include "utils.hpp"
 
 // Structure of a vertex with SetVertexShade FVF set to D3DFVF_TEX1 | D3DFVF_XYZRWH
 struct VertexTex1Xyzrwh
@@ -60,10 +31,10 @@ struct VertexTex1DiffuseXyz
 DIFFABLE_STATIC(VertexTex1Xyzrwh, g_PrimitivesToDrawVertexBuf[4]);
 DIFFABLE_STATIC(VertexTex1DiffuseXyzrwh, g_PrimitivesToDrawNoVertexBuf[4]);
 
-AnmManager::AnmManager() : virtualMachine()
+AnmManager::AnmManager()
 {
     memset(this, 0, sizeof(AnmManager));
-    for (size_t i = 0; i < sizeof(this->sprites) / sizeof(this->sprites[0]); i++)
+    for (int i = 0; i < ARRAY_SIZE_SIGNED(this->sprites); i++)
     {
         this->sprites[i].sourceFileIndex = -1;
     }
